infer/requirements.txt
pillow
requests
numpy
Flask
opencv-python 
albumentations 
numpy
fastprogress 
tqdm 
matplotlib 
torchvision
torchsummary
pycocotools 
scipy 
yacs 
loguru
